requirements.txt
streamlit
sentence-transformers
transformers
faiss-cpu
pandas
requests
numpy
torch
